float改BigDecimal,保留精确的小数

+++ src/main/scala/com/jeff/financing/dto/FlowItem.scala
@@ -1,5 +1,5 @@
 package com.jeff.financing.dto
 
 case class FlowItem(id: String, platform: Option[String], category: String,
-                    state: String, amount: Float, rate: Option[Float], dailyIncome: Option[Float],
+                    state: String, amount: BigDecimal, rate: Option[BigDecimal], dailyIncome: Option[BigDecimal],
                     target: String, startTime: Option[Long], endTime: Option[Long], createTime: Long)
